Controller-connection evaluation for the SNES and foundations for testing into a GBA.
Implemented:
 -  Controller-connection evaluation into SNESController.
 - VDebug, handles debugging into external serial monitors instead of VString.
 - GameBoyAdvance, handles mapping of the SNES pressed buttons and sending input data to the console.

Modified:
 - Now using uint16_t instead of int for pressed buttons.

+++ lib/SNESController/SNESController.h
@@ -28,8 +28,8 @@ public:
     SNESController();
     SNESController(int _clockPin, int _latchPin, int _dataPin);
 
-    int getPressedButtons() { return pressedButtons; }
-    int getPreviousPressedButtons() { return previousPressedButtons; }
+    uint16_t getPressedButtons() { return pressedButtons; }
+    uint16_t getPreviousPressedButtons() { return previousPressedButtons; }
     bool getConnected() { return connected; }
     bool getPreviousConnected() { return previousConnected; }
 
@@ -42,8 +42,8 @@ private:
     int clockPin;
     int latchPin;
     int dataPin;
-    int pressedButtons;
-    int previousPressedButtons;
+    uint16_t pressedButtons;
+    uint16_t previousPressedButtons;
     bool connected;
     bool previousConnected;
